Nulls now can be part of the CSV data
This solves an issue in the original author's implementation which required a special EOF indicator. I used Null as this indicator, but this broke other things.
The UTF-* characters that cannot be translated into LSMBC get converted to Null when handled by Lotus Script, and this broke the processing if Null is used as a special character.
This update fixes the problem.